+++ src/app/project/[slug]/setting/page.tsx
@@ -7,7 +7,7 @@ import ProjectSettingCrewAuth from '@/components/project/setting/crewAuth/Projec
 import ProjectSettingEndProject from '@/components/project/setting/endProject/ProjectSettingEndProject';
 import ErrorMessage from '@/components/ui/error/ErrorMessage';
 import ErrorPageContainer from '@/components/ui/error/ErrorPageContainer';
-import { useProjectManageAuth } from '@/features/project/auth/global/service/getProjectManageAuth';
+import { useProjectManageAuth } from '@/features/project/auth/projectManageAuth/service/getProjectManageAuth';
 
 function SettingPage({
   searchParams: { projectId },
