+++ src/components/project/setting/crewAuth/CrewAuthRow.tsx
@@ -10,7 +10,7 @@ import {
 import { numStrToBigInt } from '@/utils/common';
 import useSnackbar from '@/hooks/common/useSnackbar';
 import CrewAuthSelector from '@/components/project/setting/crewAuth/CrewAuthSelector';
-import { ProjectAuthMap } from '@/features/project/auth/global/types/projectAuth';
+import { ProjectAuthMap } from '@/features/project/auth/projectManageAuth/types/projectAuth';
 import { ProjectCrew } from '@/features/project/auth/crews/types';
 
 function CrewAuthRow({
