+++ src/components/project/setting/crewAuth/ProjectSettingCrewAuth.tsx
@@ -3,7 +3,7 @@ import SettingTitle from '@/components/project/setting/SettingTitle';
 import useProjectCrewList from '@/hooks/project/crew/useProjectCrewList';
 import ProjectSettingCrewAuthSkeleton from '@/components/project/setting/crewAuth/ProjectSettingCrewAuthSkeleton';
 import CrewAuthRow from '@/components/project/setting/crewAuth/CrewAuthRow';
-import { useProjectManageAuth } from '@/features/project/auth/global/service/getProjectManageAuth';
+import { useProjectManageAuth } from '@/features/project/auth/projectManageAuth/service/getProjectManageAuth';
 
 function ProjectSettingCrewAuth({ projectId }: { projectId: string }) {
   const { crewList, isFetching } = useProjectCrewList(projectId);
