+++ src/components/project/setting/info/ProjectSettingInfo.tsx
@@ -18,7 +18,7 @@ import SettingContainer from '@/components/project/setting/SettingContainer';
 import SettingTitle from '@/components/project/setting/SettingTitle';
 import SettingBody from '@/components/project/setting/SettingBody';
 import { bigIntToString } from '@/utils/common';
-import { ProjectAuthMap } from '@/features/project/auth/global/types/projectAuth';
+import { ProjectAuthMap } from '@/features/project/auth/projectManageAuth/types/projectAuth';
 
 function ProjectSettingInfo({
   projectId,
