+++ src/features/project/auth/leaveProject/service/leaveProject.ts
@@ -1,5 +1,5 @@
 import { request } from '@/lib/clientApi/request';
-import { ProjectAuthCode } from '@/features/project/auth/global/types/projectAuth';
+import { ProjectAuthCode } from '@/features/project/auth/projectManageAuth/types/projectAuth';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { ApiResult, ResponseBody } from '@/utils/type';
 import { CREW_NOTICE_LIST_QUERY_KEY } from '@/features/project/auth/notice/service/getCrewNoticeList';
