+++ src/features/project/auth/projectCrews/store/CrewFWModalStateStore.tsx
@@ -1,7 +1,7 @@
 import { ModalState } from '@/utils/type';
 import { atom } from 'recoil';
-import { ProjectAuthCode } from '@/features/project/auth/global/types/projectAuth';
-import { PROJECT_AUTH_CODE } from '@/features/project/auth/global/constants/projectAuthCode';
+import { ProjectAuthCode } from '@/features/project/auth/projectManageAuth/types/projectAuth';
+import { PROJECT_AUTH_CODE } from '@/features/project/auth/projectManageAuth/constants/projectAuthCode';
 import { CreateCrewFWVoteReason } from '@/features/project/auth/vote/service/createFWVote';
 
 export type CrewFWModalState = ModalState & {
