+++ src/features/project/auth/projectCrews/types.ts
@@ -1,4 +1,4 @@
-import { ProjectAuthMap } from '@/features/project/auth/global/types/projectAuth';
+import { ProjectAuthMap } from '@/features/project/auth/projectManageAuth/types/projectAuth';
 import { Position, ProjectUser, User } from '@/utils/type';
 import { CREW_STATUS } from '@/features/project/auth/crews/constants/crewStatus';
 import { TaskPointType } from '@/features/project/auth/global/types/TaskPointType';
