+++ src/features/project/auth/projectJobs/service/milestone/deleteMilestone.ts
@@ -2,7 +2,7 @@ import { ApiResult, ResponseBody } from '@/utils/type';
 import { request } from '@/lib/clientApi/request';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { getMilestonesQueryKey } from '@/features/project/auth/jobs/service/milestone/getMilestones';
-import { ProjectAuthCode } from '@/features/project/auth/global/types/projectAuth';
+import { ProjectAuthCode } from '@/features/project/auth/projectManageAuth/types/projectAuth';
 
 export type DeleteMilestoneReqData = {
   milestoneId: bigint;
