+++ src/features/project/auth/projectJobs/service/milestone/updateMilestone.ts
@@ -3,7 +3,7 @@ import { ApiResult, ResponseBody } from '@/utils/type';
 import { request } from '@/lib/clientApi/request';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { getMilestonesQueryKey } from '@/features/project/auth/jobs/service/milestone/getMilestones';
-import { ProjectAuthCode } from '@/features/project/auth/global/types/projectAuth';
+import { ProjectAuthCode } from '@/features/project/auth/projectManageAuth/types/projectAuth';
 
 export const updateMilestoneSchema = z.object({
   content: z.string().nonempty('마일스톤 내용을 입력해 주세요'),
