+++ src/features/project/auth/projectJobs/service/task/deleteTask.ts
@@ -2,7 +2,7 @@ import { request } from '@/lib/clientApi/request';
 import { ApiResult, ResponseBody } from '@/utils/type';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { getTaskListQueryKey } from '@/features/project/auth/jobs/service/task/getTaskList';
-import { ProjectAuthCode } from '@/features/project/auth/global/types/projectAuth';
+import { ProjectAuthCode } from '@/features/project/auth/projectManageAuth/types/projectAuth';
 
 export type TaskDeleteReqData = {
   workId: bigint;
