+++ src/features/project/auth/projectVote/service/createFWVote.ts
@@ -1,5 +1,5 @@
 import { request } from '@/lib/clientApi/request';
-import { ProjectAuthCode } from '@/features/project/auth/global/types/projectAuth';
+import { ProjectAuthCode } from '@/features/project/auth/projectManageAuth/types/projectAuth';
 import { useMutation } from '@tanstack/react-query';
 import { ApiResult, ResponseBody } from '@/utils/type';
 import { z } from 'zod';
