+++ src/service/project/setting/crewAuth.ts
@@ -1,7 +1,7 @@
 import { request } from '@/lib/clientApi/request';
 import { isEqual } from 'lodash';
 import { throwErrorIfInvalid } from '@/utils/common';
-import { ProjectAuthCode } from '@/features/project/auth/global/types/projectAuth';
+import { ProjectAuthCode } from '@/features/project/auth/projectManageAuth/types/projectAuth';
 
 export type ProjectSettingCrewAuthUpdData = {
   authMap: ProjectAuthCode;
